OIDC multi-tenant and lazy loading implementation
