+++ security/providers/idcs-mapper/src/main/java/io/helidon/security/providers/idcs/mapper/IdcsRoleMapperRxProviderBase.java
@@ -101,6 +101,7 @@ public abstract class IdcsRoleMapperRxProviderBase implements SubjectMappingProv
      */
     protected IdcsRoleMapperRxProviderBase(Builder<?> builder) {
         this.oidcConfig = builder.oidcConfig;
+        this.oidcConfig.tokenEndpointUri(); //Remove once IDCS is rewritten to be lazily loaded
         this.defaultIdcsSubjectType = builder.defaultIdcsSubjectType;
         if (builder.supportedTypes.isEmpty()) {
             this.supportedTypes.add(SubjectType.USER);
