+++ security/providers/oidc-common/src/main/java/module-info.java
@@ -43,4 +43,5 @@ module io.helidon.security.providers.oidc.common {
     requires java.ws.rs;
 
     exports io.helidon.security.providers.oidc.common;
+    exports io.helidon.security.providers.oidc.common.spi;
 }
